Bug/file download
This changes the behavior of file fields so that the associated file isn't loaded into memory and only a filename/URI is returned instead. This really helped speed up listing/querying the Flight model. The file can still be accessed by explicitly calling the `download_using_file_name` method on the client